Actualización de codigo

+++ src/main/java/com/platzimarket/persistence/entity/Cliente.java
@@ -19,7 +19,7 @@ public class Cliente {
     private String apellidos;
 
     @Column(name = "CELULAR")
-    private Integer celular;
+    private Long celular;
 
     @Column(name = "DIRECCION")
     private String direccion;
@@ -62,11 +62,11 @@ public class Cliente {
         this.apellidos = apellidos;
     }
 
-    public Integer getCelular() {
+    public Long getCelular() {
         return celular;
     }
 
-    public void setCelular(Integer celular) {
+    public void setCelular(Long celular) {
         this.celular = celular;
     }
