+++ src/main/java/com/platzimarket/persistence/entity/Producto.java
@@ -15,7 +15,7 @@ public class Producto {
     private String nombre;
 
     @Column(name="ID_CATEGORIA")
-    private Integer idCategoria;
+    private Long idCategoria;
 
     @Column(name="CODIGO_BARRAS")
     private String codigoBarras;
@@ -57,11 +57,11 @@ public class Producto {
         this.nombre = nombre;
     }
 
-    public Integer getIdCategoria() {
+    public Long getIdCategoria() {
         return idCategoria;
     }
 
-    public void setIdCategoria(Integer idCategoria) {
+    public void setIdCategoria(Long idCategoria) {
         this.idCategoria = idCategoria;
     }
